fix: Move utils method to existing class
