--- dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/util/StringUtils.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 2004-2025, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.util;
-
-import javax.annotation.Nonnull;
-
-public class StringUtils {
-
-  private StringUtils() {}
-
-  /**
-   * Replace all matchingChars with replaceWith chars recursively. See {@link
-   * StringUtilsTest#stringCharsReplaceTest} for examples.
-   *
-   * @param string string to operate on
-   * @param matchingChars matching chars to be replaced
-   * @param replaceWith chars to replace matching chars
-   * @return potentially-updated string
-   */
-  public static String replaceAllRecursively(
-      @Nonnull String string, @Nonnull String matchingChars, @Nonnull String replaceWith) {
-    if (!string.contains(matchingChars)) return string;
-    return replaceAllRecursively(
-        string.replace(matchingChars, replaceWith), matchingChars, replaceWith);
-  }
-}
