--- dhis-2/dhis-services/dhis-service-core/src/test/java/org/hisp/dhis/util/StringUtilsTest.java
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) 2004-2025, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.util;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.util.stream.Stream;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-
-class StringUtilsTest {
-
-  @ParameterizedTest
-  @MethodSource("stringReplaceParams")
-  @DisplayName("String char replacement should have expected outcome")
-  void stringCharsReplaceTest(
-      String original, String matchingChars, String replaceWith, String expectedString) {
-    assertEquals(
-        expectedString, StringUtils.replaceAllRecursively(original, matchingChars, replaceWith));
-  }
-
-  private static Stream<Arguments> stringReplaceParams() {
-    return Stream.of(
-        Arguments.of("test/this/string", "//", "/", "test/this/string"),
-        Arguments.of("test//this/string", "//", "/", "test/this/string"),
-        Arguments.of("test//this//string", "//", "/", "test/this/string"),
-        Arguments.of("/test//this//string", "//", "/", "/test/this/string"),
-        Arguments.of("//test/this/string//", "//", "/", "/test/this/string/"),
-        Arguments.of("///////test/////this/string////", "//", "/", "/test/this/string/"),
-        Arguments.of("////test//this////string", "//", "/", "/test/this/string"),
-        Arguments.of("test/this/string///////", "//", "/", "test/this/string/"));
-  }
-}
