+++ dhis-2/dhis-support/dhis-support-commons/pom.xml
@@ -105,6 +105,10 @@
       <groupId>org.locationtech.jts</groupId>
       <artifactId>jts-core</artifactId>
     </dependency>
+    <dependency>
+      <groupId>com.google.code.findbugs</groupId>
+      <artifactId>jsr305</artifactId>
+    </dependency>
 
     <!-- Test -->
     <dependency>
