+++ dhis-2/dhis-support/dhis-support-commons/src/main/java/org/hisp/dhis/commons/util/TextUtils.java
@@ -670,4 +670,20 @@ public class TextUtils {
   public static Set<String> getVariableNames(String input) {
     return RegexUtils.getMatches(VARIABLE_PATTERN, input, 1);
   }
+
+  /**
+   * Replace all matchingChars with replaceWith chars recursively. See {@link
+   * TextUtilsTest#stringCharsReplaceTest} for examples.
+   *
+   * @param string string to operate on
+   * @param matchingChars matching chars to be replaced
+   * @param replaceWith chars to replace matching chars
+   * @return potentially-updated string
+   */
+  public static String replaceAllRecursively(
+      @Nonnull String string, @Nonnull String matchingChars, @Nonnull String replaceWith) {
+    if (!string.contains(matchingChars)) return string;
+    return replaceAllRecursively(
+        string.replace(matchingChars, replaceWith), matchingChars, replaceWith);
+  }
 }
