+++ dhis-2/dhis-support/dhis-support-commons/src/test/java/org/hisp/dhis/commons/util/TextUtilsTest.java
@@ -40,8 +40,13 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Stream;
 import org.hisp.dhis.util.MapBuilder;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 /**
  * @author Lars Helge Overland
@@ -301,5 +306,26 @@ class TextUtilsTest {
   @Test
   void testGetVariableNamesWithNullInput() {
     assertEquals(Set.of(), TextUtils.getVariableNames(null));
+  }
+
+  @ParameterizedTest
+  @MethodSource("stringReplaceParams")
+  @DisplayName("String char replacement should have expected outcome")
+  void stringCharsReplaceTest(
+      String original, String matchingChars, String replaceWith, String expectedString) {
+    assertEquals(
+        expectedString, TextUtils.replaceAllRecursively(original, matchingChars, replaceWith));
+  }
+
+  private static Stream<Arguments> stringReplaceParams() {
+    return Stream.of(
+        Arguments.of("test/this/string", "//", "/", "test/this/string"),
+        Arguments.of("test//this/string", "//", "/", "test/this/string"),
+        Arguments.of("test//this//string", "//", "/", "test/this/string"),
+        Arguments.of("/test//this//string", "//", "/", "/test/this/string"),
+        Arguments.of("//test/this/string//", "//", "/", "/test/this/string/"),
+        Arguments.of("///////test/////this/string////", "//", "/", "/test/this/string/"),
+        Arguments.of("////test//this////string", "//", "/", "/test/this/string"),
+        Arguments.of("test/this/string///////", "//", "/", "test/this/string/"));
   }
 }
